Create cde-update-images-1231h3.sh (#6)
This is a script customer has to use to update images for 1.23.1-h3 hotfix build.
The script updates livy runtime, livy spark, and spark runtime image tags in kubernetes configmaps and deployments for a virtual cluster.
The customer has to provide the Virtual cluster id dex-app-<appid> and path to cluster kubeconfig path-to-kubeconfig as parameters.
Ex:

./cde-update-images-1231h3.sh dex-app-<appid> <path-to-your-cluster-kubeconfig>